wip: refactor for dep cycles - leave Restart() unimplemented
