+++ pkg/listeners/listeners.go
@@ -27,10 +27,6 @@ import (
 
 	// Merlin
 	"github.com/Ne0nd0g/merlin/pkg/servers"
-	"github.com/Ne0nd0g/merlin/pkg/servers/http"
-	"github.com/Ne0nd0g/merlin/pkg/servers/http2"
-	"github.com/Ne0nd0g/merlin/pkg/servers/http3"
-	"github.com/Ne0nd0g/merlin/pkg/servers/tcp"
 )
 
 type BaseHandler interface {
@@ -48,9 +44,6 @@ type Encrypter interface {
 	Decrpyt()
 }
 
-// Listeners contains all of the instantiated Listener objects
-var Listeners = make(map[uuid.UUID]*Listener)
-
 // Listener is a structure for created Merlin listener with an embedded Server object
 type Listener struct {
 	ID          uuid.UUID               // Unique identifier for the Listener object
@@ -60,44 +53,19 @@ type Listener struct {
 }
 
 // New instantiates a Listener object
-func New(options map[string]string) (*Listener, error) {
+func New(server servers.ServerInterface, options map[string]string) (*Listener, error) {
 	var listener Listener
-	var err error
 
 	// Ensure a listener name was provided
 	listener.Name = options["Name"]
 	if listener.Name == "" {
 		return &listener, fmt.Errorf("a listener name must be provided")
 	}
-	// Ensure a listener with this name does not exist
-	if Exists(listener.Name) {
-		return &listener, fmt.Errorf("a listener with this name already exists")
-	}
 
 	// Get a new server object for the listener
-	switch strings.ToLower(options["Protocol"]) {
-	case "":
-		return &listener, fmt.Errorf("a listener protocol must be provided")
-	case "http", "https", "http2":
-		listener.Server, err = http.New(options)
-	case "h2c":
-		listener.Server, err = http2.New(options)
-	case "http3":
-		listener.Server, err = http3.New(options)
-	case "tcp":
-		listener.Server, err = tcp.New(options)
-	default:
-		err = fmt.Errorf("invalid listener server type: %s", options["Protocol"])
-	}
-
-	if err != nil {
-		return &listener, err
-	}
-
 	listener.ID = uuid.NewV4()
+	listener.Server = server
 	listener.Description = options["Description"]
-
-	Listeners[listener.ID] = &listener
 
 	return &listener, nil
 }
@@ -113,27 +81,27 @@ func (l *Listener) GetConfiguredOptions() map[string]string {
 
 // Restart creates a new server instance because http servers can not be reused after they are stopped
 func (l *Listener) Restart(options map[string]string) error {
-	var err error
-
-	// Stop the running instance
-	if l.Server.Status() == servers.Running {
-		if err = l.Server.Stop(); err != nil {
-			return err
-		}
-	}
-
-	// Create a new instance
-	switch l.Server.GetProtocol() {
-	case servers.HTTP, servers.HTTPS, servers.HTTP2:
-		l.Server, err = http.Renew(l.Server.GetContext(), options)
-	case servers.H2C:
-		l.Server, err = http2.Renew(l.Server.GetContext(), options)
-	case servers.HTTP3:
-		l.Server, err = http3.Renew(l.Server.GetContext(), options)
-	default:
-		err = fmt.Errorf("invalid server protocol: %d (%s)", l.Server.GetProtocol(), servers.GetProtocol(l.Server.GetProtocol()))
-	}
-	return err
+	//var err error
+	//
+	//// Stop the running instance
+	//if l.Server.Status() == servers.Running {
+	//	if err = l.Server.Stop(); err != nil {
+	//		return err
+	//	}
+	//}
+	//
+	//// Create a new instance
+	//switch l.Server.GetProtocol() {
+	//case servers.HTTP, servers.HTTPS, servers.HTTP2:
+	//	l.Server, err = http.Renew(l.Server.GetContext(), options)
+	//case servers.H2C:
+	//	l.Server, err = http2.Renew(l.Server.GetContext(), options)
+	//case servers.HTTP3:
+	//	l.Server, err = http3.Renew(l.Server.GetContext(), options)
+	//default:
+	//	err = fmt.Errorf("invalid server protocol: %d (%s)", l.Server.GetProtocol(), servers.GetProtocol(l.Server.GetProtocol()))
+	//}
+	return l.Server.Restart(options)
 }
 
 // SetOption sets the value for a configurable option on the Listener
@@ -147,134 +115,4 @@ func (l *Listener) SetOption(option string, value string) error {
 		return l.Server.SetOption(option, value)
 	}
 	return nil
-}
-
-// GetList returns a list of Listeners that exist and is used for command line tab completion
-func GetList() func(string) []string {
-	return func(line string) []string {
-		listeners := make([]string, 0)
-		for listenerUUID := range Listeners {
-			listeners = append(listeners, Listeners[listenerUUID].Name)
-		}
-		return listeners
-	}
-}
-
-// GetListenerByID finds and returns a pointer to an instantiated listener object by its ID (UUIDv4)
-func GetListenerByID(id uuid.UUID) (*Listener, error) {
-	l, exists := Listeners[id]
-	if !exists {
-		return &Listener{}, fmt.Errorf(fmt.Sprintf("a listener with an ID of %s does not exist", id))
-	}
-	return l, nil
-}
-
-// GetListenerByName finds and returns a pointer to an instantiated listener object by its name (string)
-func GetListenerByName(name string) (*Listener, error) {
-
-	if !Exists(name) {
-		return &Listener{}, fmt.Errorf("%s listener does not exist", name)
-	}
-	var listener *Listener
-	for k, v := range Listeners {
-		if name == v.Name {
-			listener = Listeners[k]
-			break
-		}
-	}
-	return listener, nil
-}
-
-// GetListenerOptions gets a map of all configurable module options
-func GetListenerOptions(protocol string) map[string]string {
-	var options map[string]string
-	switch strings.ToLower(protocol) {
-	case "http", "https", "http2":
-		options = http.GetOptions(strings.ToLower(protocol))
-	case "h2c":
-		options = http2.GetOptions()
-	case "http3":
-		options = http3.GetOptions()
-	case "tcp":
-		options = tcp.GetOptions()
-	default:
-		options = make(map[string]string)
-	}
-	options["Name"] = "Default"
-	options["Description"] = "Default listener"
-	return options
-}
-
-// GetListenerOptionsCompleter gets an array of listener options to be used for tab completion for CLI tab completion
-func GetListenerOptionsCompleter(protocol string) func(string) []string {
-	return func(line string) []string {
-		var serverOptions map[string]string
-		options := make([]string, 0)
-		switch strings.ToLower(protocol) {
-		case "http", "https", "http2":
-			serverOptions = http.GetOptions(strings.ToLower(protocol))
-		case "h2c":
-			serverOptions = http2.GetOptions()
-		case "http3":
-			serverOptions = http3.GetOptions()
-		case "tcp":
-			serverOptions = tcp.GetOptions()
-		default:
-			serverOptions = make(map[string]string)
-		}
-		for k := range serverOptions {
-			options = append(options, k)
-		}
-		options = append(options, "Name")
-		options = append(options, "Description")
-		return options
-	}
-}
-
-// GetListenerTypesCompleter returns a list of listener types that Merlin supports for CLI tab completion
-func GetListenerTypesCompleter() func(string) []string {
-	return func(line string) []string {
-		return GetListenerTypes()
-	}
-}
-
-// GetListenerTypes returns a list of listener types that Merlin supports
-func GetListenerTypes() []string {
-	var t []string
-	for v := range servers.RegisteredServers {
-		t = append(t, v)
-	}
-	return t
-}
-
-// Exists determines if the Listener has already been instantiated
-func Exists(name string) bool {
-	for _, v := range Listeners {
-		if name == v.Name {
-			return true
-		}
-	}
-	return false
-}
-
-// RemoveByID deletes a Listener from the global list of Listeners by the input UUID
-func RemoveByID(id uuid.UUID) error {
-	if _, ok := Listeners[id]; ok {
-		err := Listeners[id].Server.Stop()
-		if err != nil {
-			return err
-		}
-		delete(Listeners, id)
-		return nil
-	}
-	return fmt.Errorf("could not remove listener: %s because it does not exist", id)
-}
-
-// GetListeners is used to return a list of Listener objects to be consumed by a client application
-func GetListeners() []Listener {
-	var listeners []Listener
-	for id := range Listeners {
-		listeners = append(listeners, *Listeners[id])
-	}
-	return listeners
 }
